pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "pgvector-rag"
version = "0.5.0"
description = "A library for interfacing with pg_vector for RAG"
license = { file = "LICENSE" }
authors = [{ name = "Gavin M. Roy", email = "gavinmroy@gmail.com" }]
readme = "README.md"
requires-python = ">=3.12"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: BSD License",
    "Natural Language :: English",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Database",
    "Topic :: Scientific/Engineering :: Artificial Intelligence"
]
dependencies = [
    "anthropic",
    "markdown-it-py",
    "openai",
    "psycopg",
    "pydantic",
    "uuid-utils"
]

[project.optional-dependencies]
dev = [
    "build",
    "coverage",
    "ruff"
]

[tool.hatch.build.targets.wheel]
include = ["pgvector_rag"]

[project.urls]
Repository = "https://github.com/gmr/pgvector-rag.git"

[tool.ruff]
line-length = 79
target-version = "py312"

[tool.ruff.lint]
select = [
    "ASYNC", # flake8-async
    "B",     # flake8-bugbear
    "BLE",   # flake8-blind-except
    "C4",    # flake8-comprehensions
    "DTZ",   # flake8-datetimez
    "E",
    "W",     # pycodestyle
    "F",     # pyflakes
    "G",     # flake8-logging-format
    "I",     # isort
    "S",     # flake8-bandit
    "T20",   # flake8-print
    "UP",    # pyupgrade
]
exclude = []

[tool.coverage.xml]
output = "build/coverage.xml"

[tool.coverage.run]
branch = true
source = ["pgvector_rag"]
command_line = "-m unittest discover tests --buffer --verbose"

[tool.pyright]
reportOptionalMemberAccess = false
reportUnknownMemberType = false
